Update per feedback
