+++ tests/acceptance/steps/common.py
@@ -16,12 +16,12 @@ from __future__ import absolute_import
 
 from behave import given
 from behave import when
-from util import create_consumer_group
-from util import create_consumer_group_with_kafka_storage
-from util import create_random_group_id
-from util import create_random_topic
-from util import initialize_kafka_offsets_topic
-from util import produce_example_msg
+from steps.util import create_consumer_group
+from steps.util import create_consumer_group_with_kafka_storage
+from steps.util import create_random_group_id
+from steps.util import create_random_topic
+from steps.util import initialize_kafka_offsets_topic
+from steps.util import produce_example_msg
 
 PRODUCED_MSG_COUNT = 82
 CONSUMED_MSG_COUNT = 39
