+++ tests/acceptance/steps/config_update.py
@@ -19,9 +19,9 @@ import time
 from behave import then
 from behave import when
 from kafka.errors import MessageSizeTooLargeError
-from util import get_cluster_config
-from util import produce_example_msg
-from util import update_topic_config
+from steps.util import get_cluster_config
+from steps.util import produce_example_msg
+from steps.util import update_topic_config
 
 from kafka_utils.util.zookeeper import ZK
