+++ tests/acceptance/steps/list_groups.py
@@ -18,12 +18,12 @@ from behave import given
 from behave import then
 from behave import when
 from six.moves import range
-from util import call_cmd
-from util import create_consumer_group
-from util import create_consumer_group_with_kafka_storage
-from util import create_random_group_id
-from util import create_random_topic
-from util import produce_example_msg
+from steps.util import call_cmd
+from steps.util import create_consumer_group
+from steps.util import create_consumer_group_with_kafka_storage
+from steps.util import create_random_group_id
+from steps.util import create_random_topic
+from steps.util import produce_example_msg
 
 
 @given('we have a set of existing consumer groups')
