+++ tests/acceptance/steps/offset_advance.py
@@ -16,8 +16,8 @@ from __future__ import absolute_import
 
 from behave import then
 from behave import when
-from util import call_cmd
-from util import get_cluster_config
+from steps.util import call_cmd
+from steps.util import get_cluster_config
 
 from kafka_utils.util.zookeeper import ZK
